# swagger.yml
swagger: "2.0"
info:
  description: "Responsible for taking in requests handling card operations and sending\
    \ them to the card provider's api."
  title: "Card Service"
  contact:
    name: "Department Of Health"
    url: "https://github.com/DepartmentOfHealth-htbhf"
    email: "dh-htbhf-team@equalexperts.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "N/A"
basePath: "/"
tags:
- name: "card-controller"
  description: "Card Controller"
paths:
  /v1/cards:
    post:
      tags:
      - "card-controller"
      summary: "Create a new card"
      operationId: "createCardUsingPOST"
      consumes:
      - "application/json"
      produces:
      - "*/*"
      parameters:
      - in: "body"
        name: "card"
        description: "Details of the new card"
        required: false
        schema:
          $ref: "#/definitions/CardDTO"
      responses:
        201:
          description: "Id and description of the new card"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
      deprecated: false
  /v1/cards/{cardId}/balance:
    get:
      tags:
      - "card-controller"
      summary: "Transfer funds to a card"
      operationId: "getBalanceUsingGET"
      produces:
      - "*/*"
      parameters:
      - name: "cardId"
        in: "path"
        description: "cardId"
        required: true
        type: "string"
      responses:
        200:
          description: "Card balance"
          schema:
            $ref: "#/definitions/CardBalance"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
      deprecated: false
  /v1/cards/{cardId}/transfer:
    post:
      tags:
      - "card-controller"
      summary: "Transfer funds to a card"
      operationId: "transferFundsUsingPOST"
      consumes:
      - "application/json"
      produces:
      - "*/*"
      parameters:
      - name: "cardId"
        in: "path"
        description: "cardId"
        required: true
        type: "string"
      - in: "body"
        name: "transferRequest"
        description: "Details of the new card"
        required: false
        schema:
          $ref: "#/definitions/TransferRequestDTO"
      responses:
        200:
          description: "Payment reference"
        201:
          description: "Created"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"
      deprecated: false
definitions:
  AddressDTO:
    type: "object"
    properties:
      addressLine1:
        type: "string"
        example: "Flat B"
        description: "First line of the address"
      addressLine2:
        type: "string"
        example: "221 Baker Street"
        description: "Second line of the address"
      postcode:
        type: "string"
        example: "AA1 1AA"
        description: "The postcode of the address."
      townOrCity:
        type: "string"
        example: "London"
        description: "Town or city of the address"
    title: "AddressDTO"
    description: "The address object"
  CardBalance:
    type: "object"
    properties:
      availableBalance:
        type: "number"
        description: "Card balance including transactions that have not yet cleared"
      ledgerBalance:
        type: "number"
        description: "Card balance only including transactions that have cleared"
    title: "CardBalance"
    description: "Cleared and uncleared balances of a card"
  CardDTO:
    type: "object"
    properties:
      address:
        description: "The address the card is registered to"
        $ref: "#/definitions/AddressDTO"
      claimId:
        type: "string"
        example: "e2bb5fcb-da0f-4f2c-8795-853e99a91209"
        description: "Unique id identifying new card request"
      dateOfBirth:
        type: "string"
        format: "date"
        example: "1985-12-30"
        description: "The date of birth, in the format YYYY-MM-DD"
      email:
        type: "string"
        example: "homer@simpson.com"
        description: "Email address of the card owner"
      firstName:
        type: "string"
        example: "Homer"
        description: "First name of the card owner"
      lastName:
        type: "string"
        example: "Simpson"
        description: "Last name of the card owner"
      mobile:
        type: "string"
        example: "07700900000"
        description: "Mobile number of the card owner"
    title: "CardDTO"
    description: "The card object"
  TransferRequestDTO:
    type: "object"
    properties:
      amount:
        type: "number"
        example: 12.4
      reference:
        type: "string"
        example: "f8b96afa-7fd1-4200-9c1a-407531036b29"
    title: "TransferRequestDTO"
    description: "Card fund transfer request object"